requirements.txt
PyQt6
PyMuPDF
pypdf
Pillow
pytesseract
reportlab
